Extract source- and target-flags from flag DSL
